requirements.txt
flask
virtualenv
mysql-connector-python